fix(api): leave only async API
